requirements.txt
flask>=1.1.0
requests
pyjwt[crypto]
